+++ ashd/pipelinev2/common.py
@@ -1,0 +1,15 @@
+import sep
+
+def get_objs(data, byteswap = False):
+    if byteswap: data = data.byteswap().newbyteorder()
+    background = sep.Background(data)
+    dsub = data - background
+    objects = sep.extract(dsub, 1.5, err=background.globalrms)
+    return (objects, dsub, background)
+
+def cut_corners(objlist, thresh=30, size=[2048, 2048]):
+    for i in objlist:
+        x = i['x']; y = i['y']
+        if not (x < thresh and y < thresh) and not (x > (size[0] - thresh) and y < thresh):
+            if not (x < thresh and y > (size[1] - thresh)) and not (x > (size[0] - thresh) and y > (size[1] - thresh)):
+                yield i
