+++ ashd/pipelinev2/global_vals.py
@@ -1,0 +1,2 @@
+MAX_TRIES = 10
+MAX_FINDINGS = 3
